week4/10_simple_pinecone_query/requirements.txt
pinecone-client==3.0.3
openai==1.101.0
python-dotenv==1.1.1
